+++ src/components/playground/inputSynthesizer/LiveDataSynthesizer.svelte
